requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.1
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
altair==5.2.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   langchain-core
    #   openai
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via -r requirements.in
blinker==1.7.0
    # via streamlit
cachetools==5.3.2
    # via streamlit
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
dataclasses-json==0.6.3
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
faiss-cpu==1.7.4
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.41
    # via streamlit
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==7.0.1
    # via streamlit
isort==5.13.2
    # via -r requirements.in
jinja2==3.1.3
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.21.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
langchain==0.1.1
    # via -r requirements.in
langchain-community==0.0.13
    # via
    #   -r requirements.in
    #   langchain
langchain-core==0.1.13
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
langchain-openai==0.0.3
    # via -r requirements.in
langsmith==0.0.83
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==5.1.0
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
marshmallow==3.20.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.3
    # via
    #   altair
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==1.8.0
    # via
    #   -r requirements.in
    #   langchain-openai
packaging==23.2
    # via
    #   altair
    #   langchain-core
    #   marshmallow
    #   streamlit
pandas==2.1.4
    # via
    #   altair
    #   streamlit
pillow==10.2.0
    # via streamlit
protobuf==4.25.2
    # via streamlit
pyarrow==14.0.2
    # via streamlit
pydantic==2.5.3
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.14.6
    # via pydantic
pydeck==0.8.1b0
    # via streamlit
pygments==2.17.2
    # via rich
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.32.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via tiktoken
requests==2.31.0
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   streamlit
    #   tiktoken
rich==13.7.0
    # via streamlit
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.25
    # via
    #   langchain
    #   langchain-community
streamlit==1.30.0
    # via -r requirements.in
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.5.2
    # via langchain-openai
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.4
    # via streamlit
tqdm==4.66.1
    # via openai
typing-extensions==4.9.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.4
    # via pandas
tzlocal==5.2
    # via streamlit
urllib3==2.1.0
    # via requests
validators==0.22.0
    # via streamlit
watchdog==3.0.0
    # via streamlit
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata
